Added new BE war, updated confs and build scripts
For some reason, it looks like tomcat 8 was giving BE issues, with some JS dependencies
not being handled correctly. We'll have to dig in to that. For now, I've reverted to
tomcat 7.0.59.

I've packaged a new tinkergraph for default usages, so we're not hitting anything Ikanow
or Orient-y. The graph only has a few nodes and edges (issue an empty search to see
the whole thing), but does have a (sensible, easy) deduplication result.